Use CoraScript DLL instead
Given that we can't register the DataLogger DLL as a DCOM, we shoud use
one that we can actually register (heeeello windows!!).

This has some advantages and some disadvantages. It is easier to have
sync operations that
will give us some sort of control, but the API is way uglier.

The reference for this, can be found at the LoggerNet Cora Script
manual.